TimezoneScript
The script will ask the user for a timezone they wish to know the time in and then run through a list of those times and present that timezones time info. I used a website to get the information I needed to understand what each command does in library. The site was geeksforgeeks "get current time in different timezone using python". I only used top section about importing and calling specific information in the print statement.